RenderManShader UIs generated from shader annotations.
Using the OSL metadata spec to define the custom UIs, as it seems like the one with the best chance of becoming widely used. Currently supporting all defined widget types and pages, but not doing popup help yet.

Also added a SectionedCompoundPlugValueWidget, which makes it easy to present the children of a CompoundPlugValueWidget in a UI broken up into sections.