Plug tooltips.
Added a GafferUI.Metadata class which is used to register descriptions for the various node and plug types. Used it to implement tooltip help for plugs, with the tooltip coming from the "parameterName.help" annotation in the case of RenderManShader nodes. We still need to register appropriate metadata for all existing nodes and generate documentation from it.