RenderManShader reloading.
RenderManShader::loadShader() may now be called passing keepExistingValues=true, to update the node status when a shader has changed on disk. This is done automatically after loading a script, so that the nodes should always be in sync with what's on disk. Also added support for min and max shader annotations which are supported for float, vector, point, normal and color shader parameters. Still need to add UI elements to expose the reloading to users.