Added NodeGraph plugContextMenuSignal() and connectionContextMenuSignal()
These allow the creation of customisable right click menus for plugs and connections in the node graph, in addition to the existing right click menus for nodes. Modified StandardNodule so that it only accepts left and middle clicks for mouse drags, making the right click available for the context menu behaviour. Fixes #122.