Fixed parameter ordering in RenderManShader UI
Parameter ordering is now based on the "ri:orderedParametersNames" shader blind data. This is currently more reliable than the order of the parameter plugs, because those aren't reordered to take account of the correct position of new parameters when a shader is reloaded. Also made SectionedCompoundPlugValueWidget support ui reordering based on the order plugs or specified in the names parameter to the constructor.

Fixes #136.