Fixed NumericPlugValueWidget update
When entering an out-of-range value in NumericPlugValueWidget, it would fail to update to show the effective clamped value, if the effective clamped value was the same as before - because the plugSetSignal was never emitted. Now we ignore plugSetSignal and always update the numeric field from the actual plug value after attempting to set it. This necessitated adding _plugConnections() to the PlugValueWidget base class, to allow the derived class to block calls to _updateFromPlug() - so that we didn't end up doing the work twice in the common case of unclamped values.