Fixed highlighting bugs.
Widgets inside a highlighted tab would incorrectly display an inherited highlight state. This could be seen when dragging a node into the node editor with either nested tabs or a nested VectorData Widget. Fixed this by making the Widget css more specific (not selecting distant children of highlighted widgets) and overriding VectorDataWidget.setHighlighted to store the highlighted state explicity on the QTableView where it could be picked up easily by the stylesheet.