Fixed MeshType::affects().
It wasn't calling the base class implementation first, which meant that the "enabled" plug wasn't getting used for dirty propagation. This manifested itself in a lack of updates in the Viewer when enabling and disabling the node. Also added a todo in SceneElementProcessor, because currently dirtiness is propagated to all the output plugs by the base class, even though we know that it's only going to actually affect the plugs being processed.

Fixes #175.