Fixed save/load of scripts containing Reference nodes.
Previously, the Reference node was allowing its referenced plugs to be serialised into the main script as well, causing them to double up on reload. It now marks them all as non-dynamic after loading to prevent this behaviour.

Also modified the serialisation for Node so that child nodes of nodes are not serialised - this is the behaviour we need for the Reference but I think it holds that most Node-derived types wouldn't want their child nodes serialising - typically they would be created on construction to perform some preprocessing of inputs and wouldn't be created by the user. The exceptions to this rule are the ScriptNode and Box, whose serialisers have been updated to ensure that their child nodes are still serialised.