Fixed paste error.
When pasting an ImageNode into a script with no ImageNodes, the image module would add the global FormatPlug onto the ScriptNode. This would then be caught by ScriptNode::paste() which would try to add it to the current selection - which would fail because the selection contains only nodes. ScriptNode::paste() now filters the newly created children, adding only nodes to the selection.

Fixes issue #251.