Fixed NumericWidget dragBegin errors.
It was returning False rather than None when it wanted to reject drag starts, and instead was starting a drag with BoolData( False ) as a result. This problem is non-existent on the C++ side because of stronger typing, but some "helpful" value converters in the Cortex bindings cause the problem on the Python side. Added some helpful warnings in the Widget documentation to hopefully avoid the same mistakes in the future.