Refactored View constructors.
Previously it was quite confusing - the inputPlug argument to the View constructor was a plug to be added to the View to receive the input, and the inputPlug arguments to the derived class constructors was a plug to be connected in to that plug. Simplified things so that making the connection is an explicit step after constructing a View, and clarified the meaning of the View::View() argument.