Improved View factory mechanism, and bound views for derivation in python.
Views may now be registered by plug name and node type rather than just by plug type. Views may now be implemented in python in a limited way - we're deliberately not allowing any virtual functions to be overridden just yet as we're not sure we can consider the API as frozen. Python views are basically limited to just installing a preprocessor of some sort.