Fixed hiding of Menu when using the search box.
Previously, the __menuActionTriggered method was getting called /after/ the menu command had returned. This meant that if the menu item popped up a modal dialogue, the menu wasn't hidden until after the dialogue had returned. Connecting directly to the signal on the menu item, /before/ connecting the command means that __menuActionTriggered is called first and the menu gets hidden at the right moment.