Fixed tab focus ordering in NodeEditor.
It's still possible to use a ListContainer in such a way that you get bad focus orders, but it's fixed enough for the purposes of the NodeEditor, which is the only place bad ordering has been reported to date. See comments in ListContainer.py for possible alternative approaches that may hold some promise, but were fruitless so far.

For the most part, fixes #107.