Improved GadgetWidget focus behaviour.
It no longer accepts focus from tab events, as that's confusing. It steals focus and releases it on mouse enter and leave events, as requested in #119. I experimented with painting a border around the GadgetWidget (viewer and graph editor) when they had the focus, but I wasn't particularly keen on the way it looked. Will await feedback before going any further with that.