Fixed size issues caused by TabbedContainer size policy.
Previously we had set the TabbedContainer size policy to Ignore, regardless of where it was being used. But really we just wanted it to stop the SplitContainers in the main layout from collapsing, and the size policy broke the sizing for the Settings and About windows. Now the SplitContainer sets the size policy to ignore for all its children, restoring the real size policy when they're removed. This keeps the nice smooth splitting behaviour in the main layout, but prevents the unwanted side effects where the other windows were created too small.