Improved undo shortcut for focussed TextWidget based PlugValueWidgets.
Because the TextWidget has its own little undo queue of text edits, it also handles the undo and redo keyboard shortcuts. This was preventing the genuine graph undo/redo shortcuts from being triggered when a NumericPlugValueWidget, PathPlugValueWidget or StringPlugValueWidget had the focus. The shortcuts are now ignored unless there are text editing undo/redo events available, and the PlugValueWidget clear the text editing undos whenever they transfer changes into the plug value. This allows undo to work correctly for both intermediate (not transferred to the plug) text edits in the TextWidget, and for edits to the plug value.