Stopped FileMenu from reusing empty scripts when opening a new one.
The problem with the reuse code path was that the "unused" script may still have some modifications made to it by callbacks registered in startup files. This means that the incoming script is being loaded into a non-vanilla ScriptNode which might cause clashes. This change is required by an upcoming commit in which a startup file adds children to the variables plug of newly created scripts.