Defined standard project variables and used them to define default display, rib and ass locations.
This is all just defined in the default startup files for the gui app. Other applications are free to define totally different conventions or no conventions at all. Startup files with precedence in the searchpath can change the values for the project variables, remove them, or define variables of their own.

Fixes #407 and #59.