Fixed bug where deleted children of Boxes weren't deselected.
In the course of doing this I introduced a new Behaviour class and an OrphanRemover behaviour, which removes nodes from a set when they lose their parent. We have several other candidates for behaviours (min/max plug constraints, input generator, etc) so we'll be fleshing out the Behaviours namespace over time.

Fixes #430.