Reordered reconnect behaviour to suit InputGenerator.
When we drag an existing connection from one nodule to another, there is always an intermediate state which isn't the desired destination state. We can either break the existing connection and then make the new one (intermediate state has no connections) or make the new connection and then break the old one (intermediate state has two connections). We originally did the former, but that ran afoul of the InputGenerator, which might remove plugs when a connection is broken, including removing the plug which is the new destination. So we now reverse that order.

Fixes #302.