Fixed TabbedContainer to hide corner widget when tabs are hidden.
The current approach might not interact well with other code which wishes to manage the visibility of the corner widget - for instance it might already be hidden when the tabs are hidden, and then get shown in an unwanted manner when the tabs are shown. The solution would be to stash the visibility when changing it due to tab visibility, but I'd prefer to wait until we have a use case before figuring out the details of that.