Added Serialiser::postScript() method.
This allows finalisation steps to be performed after all plug connections have been made. The Shader and ParameterisedHolder serialisers now use it to load their contents after the scene has been set up. Prior to this they were loading before the plug connections were being made, making the loading/upgrade steps harder to code and more error prone.

Now the loading/upgrading is performed on a completely set up scene, the process is identical to when a shader is reloaded or a parameterised object upgraded interactively by the user.