Added NodeEditor.nodeUI() method.
This is a little trickier than being a direct accessor to self.__nodeUI because of the mechanism for only updating the ui lazily. The new NodeEditor._doPendingUpdate() method makes it pretty straightforward though - I think it's best for this logic to be in the base class in case other derived classes have similar needs.