Added dynamic requirement plugs to Executable.
This makes a new plug available each time the last is used, using an ArrayPlug. This also meant fixing a bug in defaultRequirements() whereby a disconnected requirement plug led to an infinite loop.
Additionally did a todo that could have been done in the first place, and fixed Plug flags which were omitting the defaults.