Added support for Parameter ui type hint "presets".
This can be used to use the PresetsOnlyParameterValueWidgets for parameters which have presets but for which presetsOnly != True. This necessitated changing the precedence of type hints within ParameterValueWidget.create(). Matching type hints for all subclasses of the parameter are now considered to be a better match than a more specific parameter type registrations lacking the type hint.