Added custom formatters and name depth control to NameLabel.
The formatters can be used to control the conversion from the name of the GraphComponent to the label text. The number of components controls how much of the GraphComponent's ancestry to display in the name. When displaying more than one level of name, the NameLabel copes properly with renaming of parents and even reparenting.