Improved Backdrop selection and resizing.
The resize handles are now a bit larger (5 pixels) and easier to hit. The top resize handle scales down as the the user zooms out, to prevent it from stopping the user from picking the title area to select the node. Also fixed a bug which meant there was no immediate redraw when an empty Backdrop was selected.