Fixed Backdrop cut+paste bug.
More specifically, backdrops were selecting their children (framed nodes) whenever they themselves were selected, regardless of the source of that selection event. When a backdrop was being pasted, it was being pasted over the top of the original, getting selected, selecting all it's children (including the ones from the original, because they were contained in the bounds) and then getting moved to the new location, along with the bogus selected nodes.

This fix moves the responsibility for propagating selection to the backdrop children to the GraphGadget, meaning the propagation only occurs on buttonPress events, and not at any other time. This does increase the coupling between the GraphGadget and Backdrop but is much better than not fixing it. You could also argue that there was implied coupling before, because the Backdrop code was assuming it knew what the GraphGadget would be doing, even if it wasn't using the GraphGadget API.

Fixes #539.