Added even-more-simplified mode to StandardNodeUI.
This mode is specified using the new Bare value, and displays the single main column without a scrollbar. This makes it particularly suitable for embedding in other UIs, and in particular the FilterPlugValueWidget is now using it to prevent the triple-scrollbar layout it used to have when being displayed very small with a PathFilter in it.

Fixes #549.