Moved ArnoldShader::parameterValue() implementation to base class.
The base class implementation makes sense alongside the base class implementation of parameterHash(), and should work for any shader types which support the linking of parameters by specifying a value of "link:shaderHandle" to the renderer. It paves the way for an easier implementation of the upcoming OSLShader.

The RenderManShader implementation of parameterValue() still isn't shared by the base class but hopefully one day it can be - at least we'll be moving in the right direction by sharing code between ArnoldShader and OSLShader.