Stopped annoying startup errors when 3delight or arnold are missing.
We now only attempt to load GafferArnold if the arnold python module is on the PYTHONPATH, and only attempt to load GafferRenderMan if the DELIGHT environment variable is set. We do still catch and report errors if loading subsequently fails due to a configuration error though. Since the conditions under which we try to load the modules are created by the steps in the getting started guide, hopefully this will work reasonably well, not making false attempts at loading and still providing hints when things aren't quite right.

Fixes #486.