Added OSLRenderer class.
This implements the IECore::Renderer interface, allowing it to receive options, attributes and shaders as for a real renderer, but it doesn't actually produce images. Instead it returns ShadingEngines which can be used to run the OSL shader network specified by the current attribute state. This will be used by upcoming nodes to run shaders using image channels and primitive variables as inputs and outputs.