Improved path matching to provide more complete information.
Previously, PathMatcher::match() returned either Match or DescendantMatch, despite the fact that it was perfectly possible for both to be true at the same time. It now returns a bitmask communicating the full information, and also adds a new AncestorMatch value for cases where an ancestor matches.

This will make it possible to fix existing problems in the Prune and SceneElementProcessor nodes, and then to subsequently write an Isolate node.