Added OSLObject node.
This allows OSL shaders to be used as deformers on objects in the scene. Exposed access to OSLImage's internal cache of ShadingEngines so they could be shared by OSLObject - did consider splitting it out into its own file but it seemed a bit fussy - maybe if functionality grows to support transforms and attributes in the ShadingEngine then we can reconsider.