Allowed PathFilter paths plug to receive inputs.
I had disallowed it because in theory you could connect an expression into it where the expression depended on scene:path, and that would mean that ancestor/descendant queries could fail. However this prevented lots of sensible use cases including plug promotion - it's preferable to allow the sensible use cases, as the problematic case is really obscure and unlikely to occur in practice

When we start implementing AttributeFilters and the like, we might add a mechanism whereby filters can specify that they can't support efficient ancestor/descendant queries, and when we detect an input from a ComputeNode the PathFilter could perhaps disable those queries. But that might be overthinking it again too...