Prevented rogue connections being made to Shader "parameters" Plug.
CompoundPlugs track the connections of their children, and connect themselves to match the child connections in the case that all children receive inputs from plugs which share the same parent. When connecting the output of an OSL node into the input of an OSL node with only a single parameter, this could lead to an unwanted connection being made to the top level parameters plug in addition to the wanted connection made to the child plug. Removing the AcceptsInputs flag on the parameters plug fixes the problem.