Added widget overlays to GLWidget.
This allows any Widget subclass to be displayed as an overlay on top of the OpenGL contents. This initial implementation provides very little in the way of layout control for the overlays, but just enough to get a toolbar working in the viewer (in an upcoming commit). There's enough subtlety to the integration of QGraphicsViews into the Widget hierarchy that I think getting this functionality in and tested before we worry too much about layouts is a good thing. The aim in the current implementation is to have overlay Widgets appear as a standard part of the Gaffer parenting hierarchy, with none of the special cases and quirks involved behind the scenes at the Qt level.