Fixed missing shader nodules.
Because we removed the parameters CompoundPlug AcceptsInput flag, the nodule creation was being skipped. We now make nodules for plugs even if they don't accept inputs - this makes sense in the shader case where the child plugs do accept inputs, but also in the case that the plug is an output. If any UIs were relying on the old behaviour, they will now need to register a null Nodule creator - the ReferenceUI change here provides an example of this.