Fixed NodeGraph framing bug.
It was framing every time a node was moved. This was happening because I moved the dropSignal connection from self to __gadgetWidget in the recent rejig "for consistency". But because the GadgetWidget generates drags for moving nodes, but doesn't handle drops, we were ending up with drops which we didn't want.