Added bookmarks support to PathParameter uis.
Parameters can specify the bookmarks category they wish to use with a ["UI"]["bookmarksCategory"] user data entry.

I'm not super happy about having to introduce the extra parameter to the _pathChooserDialogueKeywords() method. The problem is that the method (and the ones above it) are called before the class has constructed fully, so the parameter can't be accessed with the methods provided by the base class. We need to check that Jabuka isn't reimplementing this method somewhere, and if it does, get it updated with the new signature. The better fix is probably to make the path
and pathChooserDialogueKeywords into arguments to the constructor so the derived classes have to pass them in - if people would prefer that and it won't cause too much disruption, I'd be happy to reimplement it that way.