Added bookmark support to op app.
This meant modifying the OpHolderDialogue to accept either an op as before, or an op holder node with the op on it. This is because the bookmarks are linked to the application, and the ui can't find that unless there's a parenting chain from the node up to the application. It seems that this kind of requirement is becoming more prevalent, so in the long term I expect everything to have to belong to a script and an application.

Fixes #55.