Fixed Timeline start/end field sizing bug.
The Timeline itself sets an explicit size for the fields, but this was getting overwritten by the NumericPlugValueWidget every time it updated the value from the plug. Since the max value of a plug is not mutable, it's sufficient for NumericPlugValueWidget to only set the width on construction and in setPlug(), leaving UIs that know better to override the width, safe in the knowledge that it will stay there.

Fixes #111.