Added DeleteAttributes node to GafferScene.
This derives from a new AttributeProcessor base class which simplifies the development of other nodes which solely process the values of certain named attributes. Also moved the bindings for all the attributes nodes into a separate file - the GafferSceneModule.cpp compile times are ridiculous, so we'll need to continue to split off logical groupings of nodes into separate functions like this until they're not so ridiculous.

Fixes #587.