Made ConfirmationDialogue.waitForConfirmation() focus the default button.
This makes things like the unsaved changes dialogue much more useable. There's a good case for having public focus methods on the Widget class to make this sort of thing possible without cheeky Qt API access, but I'm still undecided on the best semantics. For most things we have set/get, but focus is a bit different in that setting focus on one Widget will take it away from another. Currently we have set/get in MultiLineTextWidget and grab in TextWidget - once we've settled on the right naming we can put something on the Widget base class instead and tidy this stuff up.